% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_grids.R
\name{calc_variogram_range}
\alias{calc_variogram_range}
\title{Determine the range of spatial autocorrelation}
\usage{
calc_variogram_range(grid, maxX)
}
\description{
Determine the range of spatial autocorrelation
}

