% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_danz_2d.R
\name{Danz2012JVegSci_2D}
\alias{Danz2012JVegSci_2D}
\title{Estimation of abruptness of the vegetation-transition along an environmental gradient of an ecotone transect}
\usage{
Danz2012JVegSci_2D(i, b, migtype, ecotoner_settings, etband, etmeasure,
  copy_FromMig1_TF, do_figures, ...)
}
\arguments{
\item{i}{An integer value. The number of the ecotone transect; a value smaller than or equal to \code{transect_N(ecotoner_settings)}.}

\item{b}{An integer value. The index of the neighborhood window size; ; a value smaller than or equal to \code{neighbors_N(ecotoner_settings)}.}

\item{migtype}{A character string; one of \code{get("migtypes", envir = etr_vars)}.}

\item{ecotoner_settings}{An object of the class \linkS4class{EcotonerSettings}.}

\item{etband}{The object of the ecotone transect according to index \code{i} that was returned by the function \code{\link[ecotoner]{detect_ecotone_transects_from_searchpoint}}.}

\item{etmeasure}{The object of the ecotone transect according to index \code{i} that is created by the function \code{\link[ecotoner]{detect_ecotone_transects_from_searchpoint}}.}

\item{copy_FromMig1_TF}{A logical value. If \code{TRUE} and \code{migtype} is not \dQuote{AllMigration}, then output is copied from \dQuote{AllMigration} instead of re-calculated.}

\item{do_figures}{A logical value. If \code{TRUE}, then figures are generated and saved to disk.}

\item{...}{Further arguments. Currently implemented are \code{seed} (to seed the random number generator), \code{flag_bfig} (character string to include in file names of figures), and \code{dir_fig} (character string of the path where figures are to be stored on disk).}
}
\value{
The object \code{etmeasure} with output of the call to \code{\link{calc_Danz2012_abruptness_2D}} based on \code{doms}, \code{b}, and \code{migtype} integrated.
}
\description{
Implementation of the methods by Danz et al. (2012).
}
\references{
Danz, N. P., L. E. Frelich, P. B. Reich, and G. J. Niemi. 2012. Do vegetation boundaries display smooth or abrupt spatial transitions along environmental gradients? Evidence from the prairie-forest biome boundary of historic Minnesota, USA. Journal of Vegetation Science 24:1129-1140.
}
\seealso{
\code{\link{calc_Danz2012_abruptness_2D}}
}

