% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotone_define.R
\name{determine_ecotone_linear_extent}
\alias{determine_ecotone_linear_extent}
\title{Definition: Ecological Boundary Zone}
\usage{
determine_ecotone_linear_extent(dens_BSE, dens_Veg1, end1_toLeft, dens_TF,
  dens_Veg2, end2_toLeft, veg_density_low, veg_density_high,
  veg_density_extended_min)
}
\description{
Definition: Ecological Boundary Zone
}

