% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_ecotoner.R
\name{dir_create}
\alias{dir_create}
\title{Creates a path}
\usage{
dir_create(path, showWarnings = FALSE, recursive = TRUE, mode = "0777")
}
\arguments{
\item{path}{a character vector containing a single path name.  Tilde
    expansion (see \code{\link{path.expand}}) is done.}

\item{showWarnings}{logical; should the warnings on failure be shown?}

\item{recursive}{logical. Should elements of the path other than the
    last be created?  If true, like the Unix command \command{mkdir -p}.}

\item{mode}{the mode to be used on Unix-alikes: it will be
    coerced by \code{\link{as.octmode}}.  For \code{Sys.chmod} it is
    recycled along \code{paths}.}
}
\value{
The function returns invisible \code{TRUE} if the path exists
or if the path could successfully (and recursively by default) be created. If the path does not exist and the path cannot
be created it throws an error with a hopefully useful message.
}
\description{
This function first checks if the path exists. If the path does not exist, it attempts to create the path using
\code{\link{dir.create}} with different default values.
}
\seealso{
\code{\link{dir.create}}
}

