% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_input.R
\docType{class}
\name{EcotonerGrids-class}
\alias{EcotonerGrids}
\alias{EcotonerGrids-class}
\title{An S4-class to represent the information of the input raster grids}
\description{
An S4-class to represent the information of the input raster grids
}
\section{Slots}{

\describe{
\item{\code{grid_env}}{A \code{\linkS4class{RasterLayer}} object. The environmental grid representing the gradient to investigate, e.g., elevation}

\item{\code{grid_veg}}{A \code{\linkS4class{RasterLayer}} object. The vegetation grid representing the vegetation including the two types that make up the ecotone/ecological boundary of consideration}

\item{\code{grid_veg1}}{A \code{\linkS4class{RasterLayer}} object. The vegetation grid representing the first vegetation type of interest.}

\item{\code{grid_veg2}}{A \code{\linkS4class{RasterLayer}} object. The vegetation grid representing the second vegetation type of interest.}

\item{\code{grid_flow}}{A \code{\linkS4class{RasterLayer}} object. A grid representing the flow path of the environmental grid. Used only if \code{transect_type == 3}}

\item{\code{grid_abut}}{A \code{\linkS4class{RasterLayer}} object. A grid marking the cells where neighboring cells are of the opposite vegetation (type).}

\item{\code{grid_aspect_mean}}{A \code{\linkS4class{RasterLayer}} object. A grid representing the mean smoothed aspect of the environmental grid. Used only if \code{transect_type == 4}}

\item{\code{grid_aspect_sd}}{A \code{\linkS4class{RasterLayer}} object. A grid representing the standard deviation of the smoothed aspect of the environmental grid. Used only if \code{transect_type == 4}}

\item{\code{specs_grid}}{A \code{\linkS4class{GridInfo}} object.}

\item{\code{df_veg}}{A \code{data.frame}. A (RAT) table of 'grid_veg' in which the \code{type@ids} contained.}
}}

