% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_settings.R
\docType{class}
\name{EcotonerSettings-class}
\alias{EcotonerSettings}
\alias{EcotonerSettings-class}
\title{An S4-class to represent the settings of an 'ecotoner' project}
\description{
An S4-class to represent the settings of an 'ecotoner' project
}
\section{Slots}{

\describe{
\item{\code{transect_type}}{An integer. This flag selects the method to locate a transect, i.e., index to \code{get("transect_types", envir = etr_vars)}. See \code{\link{package?ecotoner}}. Default is \code{4} corresponding to the method by 'sufficient area of homogeneous aspect'.}

\item{\code{transect_type_desc}}{A character string. A string describing the method selected by \code{transect_type}. One of \code{get("transect_types", envir = etr_vars)}.}

\item{\code{candidate_THAs_N}}{An integer. The number of candidate transects that should be created if \code{transect_type == 4}. Default is \code{20}.}

\item{\code{transect_azimuth}}{A numeric. The orientation of the environmental gradient within a transect. Defaults to \code{3 / 2 * pi}, i.e., "left is low & right is high".}

\item{\code{cores_N}}{An integer. This project is attempted to be run on the local machine in parallel with \code{cores_N} cores, but not with more than the actual number of cores of this machine.}

\item{\code{rng_seed}}{An integer. This is the seed to set the (global) random number generator if used.}

\item{\code{reproducible}}{A logical. If \code{TRUE} then this flag sets the transect number as the seed of the random number generator for each transect function call.}

\item{\code{searchpoints_N}}{An integer. The target number of search points to randomly sample.}

\item{\code{inhibit_searchpoints}}{A logical. If \code{TRUE} then search points to initiate a transect search are sampled by a random simple sequential inhibition process; if \code{FALSE}, then a random (stratified) process is used to sample search points.}

\item{\code{transect_N}}{An integer. The number of search points generated to initiate a transect search in the neighborhood of size(s) \code{neighborhoods}.}

\item{\code{neighborhoods}}{An integer vector. The number(s) of cells per side of a square representing local grid window(s) around a search point within which the transect methods searches for a suitable transect. Neighborhood(s) must be odd. Note: Below a neighborhood of 333 cells, the density function may become unreliable.}

\item{\code{neighbors_N}}{An integer. The number of \code{neighborhoods}.}

\item{\code{stepsHullEdge}}{An integer vector. The step length(s), in grid cells, of the hull edge method by Gastner et al.}

\item{\code{clumpAreaClasses_m2}}{A numeric vector of length \code{2}. The first number is the limit between small and medium patches, and the second number is the limit between medium and large patches.}

\item{\code{res_rotate_grid}}{An integer. This number is used in the call to extract_tband_grids(); it represents variable 'fact' in raster::disaggregate(). The rotation of the grid from geographical to transect orientation requires a high resolution in cells because not all resulting cells would otherwise be represented well after the rotation. Instead, they would need to be interpolated which is tricky with categorical data such as vegetation cover. Hence, the results, e.g., patches, flowpaths, are relatively sensitive to this factor. \code{2} produced only a coarse representation of the rotated grids; \code{4} improved the resolution after rotation; \code{6} and \code{8} produced mostly similar results, but only \code{8} was satisfactorily. Default is thus \code{8}.}

\item{\code{Aspect_SDof201Mean_Maximum}}{A numeric. If \code{transect_type == 4}, then this option limits the search area for line-starting points to areas of low standard variation of the mean aspect; default is \code{75 * pi / 180 = 1.31}}

\item{\code{Aspect_DeviationAlongCandidateLines_Maximum}}{A numeric. If \code{transect_type == 4}, then this option sets the limit by which the mean aspect along candidate lines may differ from the mean aspect at the starting point differs; default is \code{30 * pi / 180 = 0.52}}

\item{\code{vegDefiningDensityTransect_low}}{A numeric. Determines the start of an ecotone transect by setting a critical vegetation density which should not be reached during the first run of sufficiently low values; default is \code{0.05}}

\item{\code{vegDefiningDensityTransect_high}}{A numeric. Determines the start of an ecotone transect by setting a critical vegetation density which should be reached during the first run of sufficiently high values; default is \code{0.75}}

\item{\code{vegDefiningDensityTransectExtended_min}}{A numeric. Determines the location of an ecotone transect by setting a critical vegetation density which should be maintained; default is \code{0.5}}

\item{\code{bandTransect_width_cellN}}{An integer. The width, in grid cell numbers, of the extracted band transect, half of which is on either side of the line transect. If \code{transect_type == 4}, then this number must match the smoothing window width of \code{\linkS4class{EcotonerGrids@grid_aspect_mean}} and \code{\linkS4class{EcotonerGrids@grid_aspect_sd}}.}

\item{\code{min_slope_with_aspect}}{A numeric. The minimum gradient, above which an aspect can be meaningfully measured/calculated. If \code{transect_type == 4}, then this number must match the used during production of \code{\linkS4class{EcotonerGrids@grid_aspect_mean}} and \code{\linkS4class{EcotonerGrids@grid_aspect_sd}}. Default is \code{2 * pi / 180 = 0.035} based for elevational slopes.}

\item{\code{type_excl}}{A \code{\linkS4class{TypeInfo}} object. This 'type' is to be excluded from 'grid_veg'.}

\item{\code{type_veg1}}{A \code{\linkS4class{TypeInfo}} object. This 'type' is one of the two vegetation (types) from 'grid_veg'.}

\item{\code{type_veg2}}{A \code{\linkS4class{TypeInfo}} object. This 'type' is one of the two vegetation (types) from 'grid_veg'.}

\item{\code{dir_prj}}{An 'EcotonerPath' object, i.e., a character string representing a path. The path to the project directory.}

\item{\code{dir_big}}{An 'EcotonerPath' object or \code{NA}. If not \code{NA}, then the path where the transect data will be stored.}

\item{\code{dir_init}}{An 'EcotonerPath' object. The path to the folder where the search points and other data are stored before transects can be searched for and extracted. This path is automatically generated.}

\item{\code{dir_out}}{An 'EcotonerPath' object. The path that contains the outputs.  This path is automatically generated from dir_prj or from dir_big if not NA.}

\item{\code{dir_out_fig}}{An 'EcotonerPath' object. The path that contains the folders where the transect figures are stored.  This path is automatically generated.}

\item{\code{dir_out_dat}}{An 'EcotonerPath' object. The path that contains the folders where the transect data are stored.  This path is automatically generated.}

\item{\code{dir_env}}{An 'EcotonerPath' object. The path to where the raster grid representing the environmental gradient 'grid_env' is stored.}

\item{\code{dir_veg}}{An 'EcotonerPath' object. The path to where the raster grid representing the vegetation 'grid_veg' is stored.}

\item{\code{dir_flow}}{An 'EcotonerPath' object. The path to where the raster grid representing the flow paths 'grid_flow' is stored (required with transect_type == 3).}

\item{\code{dir_abut}}{An 'EcotonerPath' object. The path to where the raster grid representing the abutting cells of vegetation types 1 and 2 'grid_abut' is stored (required with transect_type == 4).}

\item{\code{dir_aspect_mean}}{An 'EcotonerPath' object. The path to where the raster grid representing the mean smoothed aspect 'grid_aspect_mean' is stored (required with transect_type == 4).}

\item{\code{dir_aspect_sd}}{An 'EcotonerPath' object. The path to where the raster grid representing the standard deviation of the smoothed aspect 'grid_aspect_sd' is stored (required with transect_type == 4).}

\item{\code{file_searchpoints}}{An 'EcotonerFile' object. The entire file path to where the search points are/will be stored.}

\item{\code{file_initwindow}}{An 'EcotonerFile' object. The entire file path to where the owin object for the search point generation with inihibition is/will be stored.}

\item{\code{file_timing_locate}}{An 'EcotonerFile' object. The entire file path to the timing of each transect location call will be stored, e.g., used by the call that removes unused temporary raster files.}

\item{\code{file_etsummary}}{An 'EcotonerFile' object. The entire file path to where the table of the transect summary are stored.}

\item{\code{file_etsummary_temp}}{An 'EcotonerFile' object. The entire file path to where temporary table of the transect summary are stored.}

\item{\code{file_etmeasure_base}}{An 'EcotonerFile' object. The entire file path (minus a flag indicating which measure type) to where temporary tables of the transect measures are stored.}
}}

