% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotone_locate.R
\name{establish_ecotone_transect}
\alias{establish_ecotone_transect}
\title{Locate ecotone, prepare ecotone line and band transect, extract relevant information}
\usage{
establish_ecotone_transect(i, b, etband, etable, ipoint, ecotoner_settings,
  ecotoner_grids, elevCropped, gapCropped, bseABUTtfCropped, asp201MeanCropped,
  asp201SDCropped, tempData = NULL, dir_fig, figBasename, verbose, do_figures,
  seed)

establish_ecotone_transect(i, b, etband, etable, ipoint, ecotoner_settings,
  ecotoner_grids, elevCropped, gapCropped, bseABUTtfCropped, asp201MeanCropped,
  asp201SDCropped, tempData = NULL, dir_fig, figBasename, verbose, do_figures,
  seed)
}
\description{
Locate ecotone, prepare ecotone line and band transect, extract relevant information

Global' patch size distribution: calculate patches on gapCropped extent: hoping that this is as close as possible to global patches for the band transect extent
local' error of aspect
y-means' error of aspect
patches' of aspects grouped into the four cardinal directions
overall' error of aspect
y/x ratios' of elevation
local' elevational error from simple linear model transect elevation gradient
local' elevational error from y-means transect elevation gradient
}

