% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotone_locate.R
\name{estimate_transect_homogeneity}
\alias{estimate_transect_homogeneity}
\title{local' error of aspect
y-means' error of aspect
patches' of aspects grouped into the four cardinal directions
overall' error of aspect
y/x ratios' of elevation
local' elevational error from simple linear model transect elevation gradient
local' elevational error from y-means transect elevation gradient}
\usage{
estimate_transect_homogeneity(ecotoner_settings, ecotoner_grids, etband,
  temp_band4, aspC, aspL)
}
\description{
local' error of aspect
y-means' error of aspect
patches' of aspects grouped into the four cardinal directions
overall' error of aspect
y/x ratios' of elevation
local' elevational error from simple linear model transect elevation gradient
local' elevational error from y-means transect elevation gradient
}

