% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_ecotoner.R
\name{fun_to_apply_foreach_transect}
\alias{fun_to_apply_foreach_transect}
\title{Create a function to apply a function to each transect either in parallel with load balancing or in a serial process depending on the core settings}
\usage{
fun_to_apply_foreach_transect(ecotoner_settings)
}
\arguments{
\item{ecotoner_settings}{An object of class 'EcotonerSettings' of which the cores_N(.) will be used.}
}
\value{
A function
}
\description{
Create a function to apply a function to each transect either in parallel with load balancing or in a serial process depending on the core settings
}

