% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_input.R
\docType{class}
\name{GridInfo-class}
\alias{GridInfo}
\alias{GridInfo-class}
\title{An S4-class describing raster grid information which all of the grids must correspond to}
\description{
An S4-class describing raster grid information which all of the grids must correspond to
}
\section{Slots}{

\describe{
\item{\code{crs}}{A \link[sp::CRS]{CRS} object.}

\item{\code{res_m}}{A numeric value. The cell size of the grids in meters. The cells must be square.}

\item{\code{longlat}}{A logical. A flag indicating whether the grids' coordinates represent geographic longitude/latitude (TRUE) or not (FALSE).}

\item{\code{origin}}{A numeric vector of length two. The origin of the grids.}

\item{\code{rotation}}{A logical. A flag indicating whether the grids are rotated (TRUE) or not (FALSE).}
}}

