% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_mathematical.R
\name{Heaviside}
\alias{Heaviside}
\title{Heaviside function.}
\usage{
Heaviside(x, H0 = 0L)
}
\arguments{
\item{x}{A numeric (vector).}

\item{H0}{Selects one of three versions of the Heaviside function. A value out of \code{{0, 0.5, 1}} corresponding to \code{Heaviside(0)}}
}
\value{
A numeric vector of the same length as \code{x} with elements out of \code{{0, 1}}, respectively \code{{0, 0.5, 1}} if \code{H0 = 0.5}.
}
\description{
The piecewise constant Heaviside/unit step function.
}
\examples{
x <- (-2):2
cbind(x, Heaviside(x, H0 = 0))
cbind(x, Heaviside(x, H0 = 0.5))
cbind(x, Heaviside(x, H0 = 1))
\dontrun{cbind(x, Heaviside(x, H0 = 0.3))}
}

