% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotone_locate.R
\name{identify_migration_patches}
\alias{identify_migration_patches}
\title{Identify suitably migrated vegetation patches of the ecotone band transect, extract relevant information}
\usage{
identify_migration_patches(i, b, ecotoner_settings, etband, etable, dir_fig,
  figBasename, verbose, do_figures, seed)
}
\description{
Identify suitably migrated vegetation patches of the ecotone band transect, extract relevant information
}

