% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_mathematical.R
\name{is_odd}
\alias{is_odd}
\title{Determine whether a number is odd.}
\usage{
is_odd(x)
}
\arguments{
\item{x}{An integer (vector).}
}
\value{
A logical: TRUE, if \code{x} is odd; FALSE, otherwise
}
\description{
Determine whether a number is odd.
}
\examples{
is_odd((-3):3)
}

