% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_transect_line.R
\name{locate_candidate_THAs}
\alias{locate_candidate_THAs}
\title{Locate candidate transect lines of (sufficient) homogeneous aspect THAs}
\usage{
locate_candidate_THAs(n, grid_gradient, max_neighborhood, asp201Mean, asp201SD,
  abuts, max_aspect_sd, max_aspect_diff, width_N, seed = NULL)
}
\arguments{
\item{n}{An integer. The size of initial random sample of starting points to locate candidate transect lines
within areas of \code{abut} and where \code{asp201SD} <= \code{max_aspect_sd}.}

\item{grid_gradient}{A \code{\linkS4class{RasterLayer}} object representing the environmental gradient, e.g., elevation.}

\item{asp201Mean}{A \code{\linkS4class{RasterLayer}} object.}

\item{asp201SD}{A \code{\linkS4class{RasterLayer}} object.}

\item{abuts}{A \code{\linkS4class{RasterLayer}} object.}

\item{max_aspect_sd}{A numeric. Limits the search area for line-starting points to where \code{asp201SD} <= \code{max_aspect_sd}.}

\item{max_aspect_diff}{A numeric. Cuts the candidate lines to areas where \code{asp201Mean} differs from the mean aspect at the starting point differs by less than \code{max_aspect_diff}.}

\item{width_N}{An integer. The width of the band transect in number of grid cells which will be put around the transect line.}

\item{seed}{An integer value, \code{NA}, or \code{NULL} to set the random number generator. \code{NULL} follows the behavior of \code{\link{set.seed}}.}
}
\description{
Locate candidate transect lines of (sufficient) homogeneous aspect THAs
}

