% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_mathematical.R
\name{majority}
\alias{majority}
\title{Determines the most common element(s)}
\usage{
majority(x, use_mode = TRUE, return_single = TRUE, na.rm = FALSE,
  seed = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{use_mode}{A logical indicating whether to use \code{\link[modeest]{modeest::mfv}} if suitable.}

\item{return_single}{A logical indicating whether a single value should return even if several elements of \code{x} are the most common. If so, then one will be picked at random.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{seed}{An integer value or \code{NULL} to set the random number generator. \code{NULL} follows the behavior of \code{\link{set.seed}}.}
}
\description{
Determines the most common element(s)
}
\examples{
x1 <- c(rep(NA, 3), 1:4)
x2 <- c(rep(1/3, 3), 1:4, rep(NA, 3))
majority(x1)
majority(x1, na.rm = TRUE)
majority(x1, return_single = FALSE, na.rm = TRUE)
majority(x1, use_mode = FALSE, return_single = FALSE, na.rm = TRUE)
majority(x1, use_mode = FALSE, return_single = TRUE, na.rm = TRUE)
majority(x1, use_mode = FALSE, return_single = FALSE, na.rm = FALSE)
majority(x2, return_single = FALSE)
}

