% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotone_measure.R
\name{measure_ecotone_per_transect}
\alias{measure_ecotone_per_transect}
\title{export}
\usage{
measure_ecotone_per_transect(i, ecotoner_settings, et_methods, verbose = TRUE,
  do_figures = TRUE)
}
\description{
export
}

