% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_GIS.R
\name{mosaic_tiles}
\alias{mosaic_tiles}
\title{Mosaic raster tiles}
\usage{
mosaic_tiles(dir_tiles, chunksize = 10L, fname_grid_ned, format, ...)
}
\arguments{
\item{dir_tiles}{A character string. The path to the directory where the folders of each tile are stored on disk.}

\item{chunksize}{An integer value. The number of tiles that are mosaic-ed per function call.}

\item{fname_grid_ned}{A character string. The path with file name (with or without file extension) of the final NED raster.}

\item{format}{A character string or missing. The file type. For possible values see \code{\link{raster::writeFormats}}.}

\item{...}{Additional arguments for writing raster files to disk as in \code{\link{raster::writeRaster}}.}
}
\value{
A RasterLayer object stored on disk as \code{fname_grid_ned}
}
\description{
This function can be used to mosaic the NED tiles downloaded from nationalmap.gov
}
\details{
If 'fname_grid_ned' already exists and agree with origin, resolution, and coordinate reference system with the tiles, then tiles are added instead of creating a new raster object.
}
\seealso{
This function calls \code{\link{raster::mosaic}}
}

