% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_ecotoner.R
\name{parLapplyLBc}
\alias{parLapplyLBc}
\title{A load balancing version of clusterApply with increased level of balancing}
\usage{
parLapplyLBc(cl = NULL, X, fun, ..., chunksize = 10L)
}
\arguments{
\item{cl}{a cluster object, created by this package or by package
    \href{{https://CRAN.R-project.org/package=snow}{\pkg{snow}}}.  If \code{NULL}, use the registered default cluster.}

\item{X}{A vector (atomic or list) for \code{parLapply} and
    \code{parSapply}, an array for \code{parApply}.}

\item{fun}{function or character string naming a function.}

\item{...}{additional arguments to pass to \code{fun} or \code{FUN}:
    beware of partial matching to earlier arguments.}

\item{chunksize}{An integer value. A suggestion on how large the number of tasks should be that are sent per job to a worker.}
}
\value{
A list of the same length as \code{X} each with the returned value of \code{fun}.
}
\description{
This function is a modified copy of \code{\link{parallel::parLapplyLB}}  (v3.2.4)
}
\details{
Bug report is submitted under \href{https://bugs.r-project.org/bugzilla3/show_bug.cgi?id=16792}{# 16792}
}
\seealso{
\code{\link{parallel::parLapplyLB}} and \code{\link{parallel::parLapply}}
}

