% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_GIS.R
\name{project_raster}
\alias{project_raster}
\title{Project a raster}
\usage{
project_raster(grid_from, fname_grid_to, res_to, crs_to, parallel_N = 1, ...)
}
\arguments{
\item{grid_from}{A RasterLayer object. The raster to be projected}

\item{fname_grid_to}{A character string. The path with file name (with or without file extension) of the resulting projected raster.}

\item{res_to}{The raster cell resolution of the target.}

\item{crs_to}{The CRS of the target.}

\item{parallel_N}{An integer value. If larger than 1, then the re-projection will use a snow cluster, see \code{\link{raster::beginCluster}} for details}

\item{...}{Additional arguments for writing raster files to disk as in \code{\link{raster::writeRaster}}.}

\item{chunksize}{An integer value or missing. Maximum number of cells to read/write in a single chunk while processing (chunk by chunk) disk based Raster* objects (see \code{\link{raster::rasterOptions}} for details)}

\item{maxmemory}{An integer value or missing. Maximum number of cells to read into memory. I.e., if a Raster* object has more than this number of cells, \code{\link{raster::canProcessInMemory}} will return \code{FALSE} (see \code{\link{raster::rasterOptions}} for details)}
}
\value{
A RasterLayer object stored on disk as \code{fname_grid_to}
}
\description{
This function may be very SLOW and memory intensive!
}
\seealso{
This function calls \code{\link{raster::projectRaster}}
}

