% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_mathematical.R
\name{root_mean_square}
\alias{RMS}
\alias{RMSE}
\alias{root_mean_square}
\title{Calculate Root Mean Square (Error) RMS(E).}
\usage{
root_mean_square(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
If \code{x} are residuals, then RSME else RMS of \code{x}
}
\description{
Calculate Root Mean Square (Error) RMS(E).
}
\examples{
x <- runif(30)
y <- 0.1 + 0.3 * x + rnorm(30, 0, 0.5)
r <- resid(lm(y ~ x))
stats::sd(r)
root_mean_square(r)
}
\seealso{
\code{\link[stats]{sd}}
}

