% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_grids.R
\name{rSSI2}
\alias{rSSI2}
\title{An implementation of a Simple Sequential Inhibition (SSI) process}
\usage{
rSSI2(r, n, win, giveup = 1000, x.init = NULL, ...)
}
\arguments{
\item{r}{
    Inhibition distance.
  }

\item{n}{
    Maximum number of points allowed. 
    If \code{n} is finite, stop when the \emph{total} number of points
    in the point pattern reaches \code{n}.
    If \code{n} is infinite (the default), stop only when
    it is apparently impossible to add any more points.
    See \bold{Details}.
  }

\item{win}{
    Window in which to simulate the pattern.
    An object of class \code{"owin"}
    or something acceptable to \code{\link{as.owin}}.
    The default window is the unit square, unless 
    \code{x.init} is specified, when the default window 
    is the window of \code{x.init}.
  }

\item{giveup}{
    Number of rejected proposals after which the algorithm should terminate.
  }

\item{x.init}{
    Optional. Initial configuration of points. A point pattern
    (object of class \code{"ppp"}).  The pattern returned by
    \code{rSSI} consists of this pattern together with the points
    added via simple sequential inhibition.  See \bold{Details}.
  }

\item{...}{Ignored.}
}
\description{
rSSI of cell-center points and avoiding spatstat. This function generates a random pattern of points selected from \code{win} all of which have at least \code{r} units distance to each other.
}
\details{
This function implements the same process as \code{\link{spatstat::rSSI}}. \code{\link{spatstat::rSSI}} relies on \code{win} being of class \code{owin}, produces an object of class \code{ppp} - both very memory-intensive classes -, and copies such objects excessively. For instance, a matrix with 75 million rows (1.1 GB memory) was translated into an object of class \code{owin} of 16.8 GB memory usage. Running \code{\link{spatstat::rSSI}} on this 'owin' object failed after the process required more than than 84 GB memory - more than was available.
}
\seealso{
\code{\link{spatstat::rSSI}} which this function simplifies
}

