% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_mathematical.R
\name{set_RNG_stream}
\alias{set_RNG_stream}
\title{Sets the .Random.seed}
\usage{
set_RNG_stream(seed)
}
\arguments{
\item{seed}{A vector appropriate for \code{\link{.Random.seed}} of the current RNG; a single integer or NULL that will be passed to set.seed(); or NA which will do no work.}
}
\description{
This is here defined to be called from a function being executed on nodes/slaves with a prepared seed from the function \code{\link{prepare_RNG_streams}}, i.e., calling this function will also set RNGkind accordingly to the first element of seed.
}
\seealso{
\code{\link{set.seed}}, \code{\link{RNGkind}}
}

