% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_transect_band.R
\name{set_stband_polygon}
\alias{set_stband_polygon}
\title{Draws a rectangular band (of #cells=width) around transect defined by two points (start, end)}
\usage{
set_stband_polygon(start, end, width_n, res_m, Longitude_WGS84 = 0)
}
\description{
Draws a rectangular band (of #cells=width) around transect defined by two points (start, end)
}

