% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_ecotoner.R
\name{simplify2result}
\alias{simplify2result}
\title{Simplify a list to a data.frame}
\usage{
simplify2result(x, showWarnings = TRUE)
}
\arguments{
\item{x}{A R object. See \code{Value}.}

\item{showWarnings}{A logical value. If \code{TRUE} then warnings are collected and printed if there are any.}
}
\value{
A data.frame if x is a list of elements with the same number of columns. If the list elements do not all have the same number of columns, then a list without the NULL elements is returned. If x is not a list, then x is returned unchanged.
}
\description{
Convert a list of elements with the same number of columns to a data.frame (discarding NULL elements) and return it unchanged otherwise.
}
\seealso{
\code{\link{simplify2array}}
}

