% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitities_mathematical.R
\name{tang_curvature}
\alias{tang_curvature}
\title{Calculate tangential curvature.}
\usage{
tang_curvature(x, ..., res_m)
}
\arguments{
\item{x}{A numeric vector of length 9, i.e., the 8 cells plus the center of a 3 x 3 square grid.}

\item{...}{Not used, but required so that \code{tang_curvature} can be used as function argument to \code{\link{raster::focal}}}
}
\value{
A numeric value in the range [‑1,1] with a sign that ensures that positive curvature equates to convex forms and negative curvature equates to concave forms
}
\description{
Calculate tangential curvature.
}
\examples{
if (require(raster, quietly = TRUE)) {
    x <- raster::raster(matrix(1:36, ncol = 6), xmn=0, xmx=1, ymn=0, ymx=1)
    tcurv <- raster::focal(x, w = 3, fun = tang_curvature, pad = TRUE, padValue = NA)
}
}

