% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_transect_band.R
\name{trim1_band_to_ecotone}
\alias{trim1_band_to_ecotone}
\title{Crop band transect to limits1_ZoneEcolBoundary}
\usage{
trim1_band_to_ecotone(stband_grids, start_icol, end_icol)
}
\description{
Crop band transect to limits1_ZoneEcolBoundary
}

