% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_settings.R
\docType{class}
\name{TypeInfo-class}
\alias{TypeInfo}
\alias{TypeInfo-class}
\title{An S4-class describing the vegetation (type) of each side of the ecotone/ecological boundary}
\description{
An S4-class describing the vegetation (type) of each side of the ecotone/ecological boundary
}
\section{Slots}{

\describe{
\item{\code{ids}}{An integer (vector). The cell values of 'grid_veg' that belong to this vegetation (type).}

\item{\code{field}}{A character string. The column name of 'df_veg' that was used to define this vegetation (type) [currently, not implemented].}

\item{\code{end_to_left}}{A logical. If \code{TRUE}, then this vegetation (type) is on the 'left' side of the ecotone/ecological boundary.}
}}

