% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_input.R
\name{valid_grid}
\alias{valid_grid}
\title{Checks if a raster grid has more than one cell.}
\usage{
valid_grid(x)
}
\value{
A logical value.
}
\description{
Checks if a raster grid has more than one cell.
}

