% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_danz_2d.R
\name{calc_Danz2012_abruptness_2D}
\alias{calc_Danz2012_abruptness_2D}
\title{Estimates the abruptness of outcome y against x}
\usage{
calc_Danz2012_abruptness_2D(x1, z1, x2, z2, end_toLeft, xlab, fname,
  seed = NULL)
}
\arguments{
\item{x}{A numeric vector. Predictor along the ecotone transect from left-right. The length of x has to correspond to the ecotone transect. For instance, distance along the transect.}

\item{z}{A numeric vector. The outcome along the ecotone transect, e.g., vegetation density.}
}
\description{
Estimates the abruptness of outcome y against x
}

